# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://gestion-hotel.test/"
        }
    ],
    "paths": {
...
        "/api/huesped": {
            "get": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Obtener lista de huéspedes paginada",
                "operationId": "bc38070d79273e5f59f3b20fba3d9f48",
                "responses": {
                    "200": {
                        "description": "Lista de huéspedes paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Crear un nuevo huésped",
                "operationId": "8fc65c7e778ac9a74965afb8d49c515a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHuespedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Huésped creado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/huesped/{id}": {
            "get": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Obtener un huésped por ID",
                "operationId": "8bbefe2f420cbcb685f5a86d651fbc0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del huésped"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Actualizar un huésped",
                "operationId": "1224634c9ef39f5e38a4433379f63b56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHuespedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Huésped actualizado correctamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Eliminar un huésped",
                "operationId": "a288a3efa8373922bb141444faaaa400",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Huésped eliminado"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
        "/api/huesped/{id}/habitacion": {
            "get": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Obtener la habitación de un huésped",
                "operationId": "41fb1038fc0cbc6aea89d0cc0dbd70fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación del huésped"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
        "/api/huesped/{idHuesped}/reservar/{idHabitacion}": {
            "post": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Reservar una habitación para un huésped",
                "operationId": "e1187a17915c3bb7d6ff759a6a8be4a1",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idHabitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Habitación reservada correctamente"
                    },
                    "400": {
                        "description": "Habitación no disponible"
                    },
                    "404": {
                        "description": "Huésped o habitación no encontrado"
                    }
                }
            }
        },
        "/api/huesped/{idHuesped}/checkout": {
            "post": {
                "tags": [
                    "Huesped"
                ],
                "summary": "Registrar el check-out del huésped y liberar la habitación",
                "operationId": "afe9bc90ad3d22e1af55dd1606c57223",
                "parameters": [
                    {
                        "name": "idHuesped",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check-out registrado correctamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreHuespedRequest": {
                "title": "Store Huesped Request",
                "description": "Datos requeridos para crear un nuevo huésped",
                "required": [
                    "nombre",
                    "apellido",
                    "dniPasaporte"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del huésped",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del huésped",
                        "type": "string"
                    },
                    "dniPasaporte": {
                        "description": "DNI o pasaporte del huésped",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PutHuespedRequest": {
                "title": "Put Huesped Request",
                "description": "Datos requeridos para actualizar un huésped",
                "required": [
                    "nombre",
                    "apellido",
                    "dniPasaporte"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del huésped",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del huésped",
                        "type": "string"
                    },
                    "dniPasaporte": {
                        "description": "DNI o pasaporte del huésped",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}